backend/requirements.txt
argon2-cffi==21.1.0
asgiref==3.7.2
boto3==1.26.158
botocore==1.29.158
braintree==4.13.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==2.0.12
cryptography==41.0.1
defusedxml==0.7.1
Django==4.2.2
django-ckeditor==6.1.0
django-cors-headers==4.1.0
django-environ==0.4.5
django-js-asset==2.0.0
django-storages==1.12.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
environ==1.0
gunicorn==20.1.0
idna==3.4
jmespath==1.0.1
Markdown==3.3.4
mutagen==1.45.1
oauthlib==3.2.2
Pillow==8.1.2
psycopg2==2.9.2
pycparser==2.21
PyJWT==2.1.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.1
s3transfer==0.6.1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
stripe==2.58.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.16
whitenoise==5.3.0
